Split off model definition from gears.py example
