+++ examples/custom/gears/gear.coffee
